#199: Move ProfileHeader specific common-logic
