+++ src/components/Dashboard/Profile/sections/ProfileHeader/common/dialogStyles.ts
