+++ src/components/Dashboard/Profile/sections/ProfileHeader/common/index.ts
@@ -3,4 +3,3 @@ export * from "./labelMaps";
 export * from "./profileHeaderStyles";
 export * from "./ProfileStatusBadge";
 export * from "./StatusRowField";
-export * from "./statusMaps";
