+++ src/components/Dashboard/Profile/sections/ProfileHeader/volunteer/ChangeEngagementStatusDialog.tsx
@@ -18,7 +18,7 @@ import {
   OptionLabel,
   OptionsContainer,
   RadioOption,
-} from "../../../common";
+} from "../common";
 import { createEngagementLabelMap, ENGAGEMENT_DESCRIPTION_KEYS } from "./constants";
 import { UseEngagementStatusDialogReturn } from "./useEngagementStatusDialog";
